requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiosqlite>=0.17.0
aioschedule>=0.5.2
pillow>=9.0.0
requests>=2.26.0
python-dateutil>=2.8.2
aiohttp>=3.8.0
asyncio>=3.4.3
colorama>=0.4.4
PyYAML>=6.0
sqlite3>=3.36.0